Refactor: move suggestion functions from demand to suggestions
